+++ src/simp.c
@@ -247,9 +247,9 @@ Tree simplify(int op, Type ty, Tree l, Tree r) {
 			break;
 
 		case CVI+F:
-			xcvtcnst(I,l->u.v.i,ty,d,(long double)l->u.v.i);
+			xcvtcnst(I,l->u.v.i,ty,d,(double)l->u.v.i);
 		case CVU+F:
-			xcvtcnst(U,l->u.v.u,ty,d,(long double)l->u.v.u);
+			xcvtcnst(U,l->u.v.u,ty,d,(double)l->u.v.u);
 			break;
 		case CVF+I:
 			xcvtcnst(F,l->u.v.d,ty,i,(long)l->u.v.d);
